feat: scan NPM project dependencies
this commit does the following:

scan_npm_project.ts is called per-project and
makes use of helper methods to get the dependencies of a project and
collate them into useful information.

npm_package_helper.ts now uses a different method for importing a file
as I cannot work out why but the previous implementation
does not return the file data before moving on.

[Ticket: FYP-18]